+++ src/ui/error.rs
@@ -6,8 +6,6 @@
 // NOTE: IoError doesn't implement Clone, PartialEq, Hash, etc.
 #[cfg(feature = "sys")]
 use crate::IoError;
-#[cfg(feature = "layout")]
-use crate::LayoutError;
 
 #[doc = crate::TAG_RESULT!()]
 /// A user-interface result.
@@ -24,11 +22,6 @@ pub enum UiError {
 
     /// The requested functionality is not supported by this number type.
     NotSupported,
-
-    /// Layout-related error.
-    #[cfg(feature = "layout")]
-    #[cfg_attr(nightly_doc, doc(cfg(feature = "layout")))]
-    Layout(LayoutError),
 
     /// An io error.
     #[cfg(feature = "sys")]
@@ -49,29 +42,17 @@ impl crate::Error for UiError {}
 
 mod core_impls {
     use super::*;
-    #[cfg(feature = "layout")]
-    use crate::LayoutError;
     use crate::impl_trait;
 
     impl_trait! { fmt::Display for UiError |self, f| {
         use UiError as E;
         match self {
-            #[cfg(feature = "layout")]
-            E::Layout(e) => write!(f, "{e:?}"),
-
             E::NotImplemented => write!(f, "Not implemented."),
             E::NotSupported => write!(f, "Not supported."),
             #[cfg(feature = "sys")]
             E::Io(e) => write!(f, "{e:?}"),
         }
     }}
-
-    #[cfg(feature = "layout")]
-    impl From<LayoutError> for UiError {
-        fn from(e: LayoutError) -> Self {
-            UiError::Layout(e)
-        }
-    }
 
     #[cfg(feature = "sys")]
     impl From<IoError> for UiError {