+++ src/ui/layout/mod.rs
@@ -5,17 +5,16 @@
 // safety
 #![cfg_attr(feature = "safe_layout", forbid(unsafe_code))]
 
-mod error;
-
 crate::items! { // structural access: _mods, _all
     #[allow(unused)]
     pub use _mods::*;
 
-    mod _mods {
-        pub use super::error::*;
+    mod _mods { #![allow(unused)]
+        // WIPZONE
     }
-    pub(super) mod _all {
+    pub(super) mod _all { #![allow(unused)]
         #[doc(inline)]
         pub use super::_mods::*;
     }
 }
+// WIPZONE