+++ DOCS/CHANGELOG.md
@@ -4,6 +4,8 @@
 =======================
 
 ### Added
+- new methods:
+  - `Linux:` `print[ln]_unchecked[_fast]`, `eprint_bytes`.
 
 ### Removed